requirements.txt
# requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
ollama>=0.1.6          # <--- NEW
tenacity>=8.2.0
rich>=13.7.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.0
spacy>=3.7.0
httpx>=0.26.0